Update psycopg2 package to psycopg2-binary in requirements
The requirements.txt file has been updated to replace the psycopg2 package with psycopg2-binary. This adjustment helps avoid issues related to the psycopg2 package, which requires additional build dependencies. psycopg2-binary can be a better option for environments where these build dependencies are not available or are difficult to install.