Update psycopg2 package to psycopg2-binary in requirements (#40)
* Update psycopg2 package to psycopg2-binary in requirements

The requirements.txt file has been updated to replace the psycopg2 package with psycopg2-binary. This adjustment helps avoid issues related to the psycopg2 package, which requires additional build dependencies. psycopg2-binary can be a better option for environments where these build dependencies are not available or are difficult to install.

* Refactor code to handle absence of pandas package

This commit makes many changes aimed at addressing situations when the pandas package is not present. These changes include removing unnecessary imports, adjusting usages of pandas in ApiSuperClass.py and WrenchLogger.py to handle cases when pandas is not installed, and altering error handling to raise an exception or log a warning in case pandas is not installed. Now, the code will still function without hard dependency on pandas, although certain functionalities are limited.

* Add pandas to GitHub workflow dependencies

Pandas package has been added to the dependencies being installed in the GitHub Actions workflow for publishing to PyPI. This change ensures that in the workflow, pandas is available and the functionalities that require it can operate without issues.

* Simplify and correct package versions in requirements.txt

The specific versions of required packages have been replaced with a slightly more flexible requirement using the tilde approximation operator, which will match the most recent minor version. The inclusion of pandas was also corrected as it was not supposed to be in the requirements.

* Add pandas to the installation requirements in Github Actions workflow

The installation requirements in the Github Actions workflow for publishing to PyPi have been updated. The package pandas has been added to the packages to be installed during the workflow execution. This ensures availability of necessary packages for correct functioning of the application.