Added S3ServiceGateway for AWS S3 interactions
Implemented S3ServiceGateway class, which manages all interactions with AWS S3. This includes uploading, downloading, moving, copying, checking existence and listing of objects in the bucket. The methods implemented use a retryable decorator to manage request failures, ensuring robustness of the application. Implemented as a singleton to ensure a single instance throughout the application.