Add Retryable decorator for retrying functions
Added a new decorator, Retryable, that can be used to automatically retry a function if it raises specified exceptions. This decorator aims to improve the robustness of the code by handling transient errors. It will retry a function up to a maximum number of retries with a delay between attempts.