Add build_return_json helper function
Created a new helper function, build_return_json, that structures a JSON response body with a given status code and message. This function also supports an optional return dictionary for additional data to be included in the response body. As a result, this improves the uniformity and consistency of HTTP responses in the application.