Add new logo to resources
The new logo has been added to the resources directory in SVG format, alongside the copyright notice and other comments. This new logo will replace the existent one and will be used across the application to provide a refreshed brand identity. The SVG format allows for better quality and scalability across different platforms and screens resolutions.