Expand search paths for .env file in ConfigurationManager
The previous logic of ConfigurationManager was updated to extend the list of potential locations for the .env file. This change aims to increase the probability of locating the .env file in less conventional directory structures by checking parent and grandparent directories. Additionally, debug logging has been added to share information about the paths being checked.