Add AWS deployment configuration to SSH tunnel
Enabled AWS deployment support by introducing 'aws_deployment' flag in the ConfigurationManager. If the 'aws_deployment' flag is true, the SSH tunnel connection won't be used even when the qa_host_check is in the secret string. This flag can be configured with environment variables or kwargs for more flexible deployment scenarios.