Add single quote decoder to JsonSerializer
This commit introduces a new decoder class, `single_quote_decoder`, to the `JsonSerializer` module. The class is for pre-processing JSON strings that use single quotes for keys and values or include Markdown block markers. This function is supposed to fix issues with decoding JSON strings from certain sources that do not strictly follow the syntax rules.