Update styling and handling in WrenchLogger module
Improved the text styling and handling in the WrenchLogger module. The colors and the brightness of the printout have been adjusted. Handling of string type data, edge cases for different formatting situations and empty starts have been properly addressed. A visual enhancement to application headers was added when colorama is imported.