V2 (#53)
* Delete Docs

* Add Sphinx documentation setup

This commit introduces a new Sphinx documentation setup for the WrenchCL project. A configuration file, make file, an index, and a batch file were created to ensure efficient documentation process. The project's documentation now supports autodoc, viewcode, and napoleon extensions for better readability.

* Update GitHub Actions workflow for Python project

This commit fine-tunes multiple sections of our GitHub Actions workflow. Some redundant comments were removed for better readability while image versions for actions were updated to the latest. A critical permission block was added to resolve workflow dispatch issues. Additionally, the workflow now includes steps for generating Sphinx documentation and deploying it to GitHub Pages.

* Add Sphinx documentation generation in GitHub Actions

A new workflow file has been added to the .github directory that automatically generates Sphinx documentation. This workflow is triggered manually and deploys the built documentation to GitHub Pages. Moreover, it operates in the latest Ubuntu environment with Python 3.11.

* Enable publishing to Py

* Update pytest command in GitHub workflow

The pytest command in the GitHub actions workflow has been updated to include the coverage report in XML format. An additional parameter is also added to the Codecov action to specify the location of the coverage report file.

* Update pytest command and Codecov upload in GitHub workflow

The pytest command has been updated to correctly target the WrenchCL application for coverage report. The steps for uploading the coverage report to Codecov within the GitHub workflow were improved to remove redundant configurations and include verbose logging for better troubleshooting.

* Update pytest run command and relocate test files

The commit modified the command to execute pytest, explicitly setting it to run on the /tests directory. Also, it relocated the test modules from resources/tests/ to the tests/ directory, improving the project structure and organization.

* Update pytest run command and relocate test files

The commit modified the command to execute pytest, explicitly setting it to run on the /tests directory. Also, it relocated the test modules from resources/tests/ to the tests/ directory, improving the project structure and organization.

* Update pytest command in GitHub workflows

The pytest command in the GitHub workflow has been simplified. The target test directory "./tests" has been placed after the "--cov=WrenchCL" argument for clarity. Additionally, the verbose mode previously used with the Codecov action is turned off to minimize output verbosity.

* Optimize pytest command and remove Codecov action

The commit simplifies the pytest command in the GitHub workflows configuration file, making the target test directory more explicit. Additionally, the Codecov