Separate documentation generation and deployment from PyPi publishing
The code related to Sphinx documentation generation and deployment to GitHub Pages has been moved from the 'publish-to-pypi' workflow to a separate workflow named 'generate-docs'. This separation makes each workflow specific to a single responsibility, simplifies the 'publish-to-pypi' workflow, and provides greater flexibility in controlling when and how documentation is built and deployed.