Add fetch for all tags in GitHub Actions
The update ensures that all Git tags are fetched before trying to get the latest version tag in the GitHub Actions workflow for generating documentation. It also updates the fall back version to 'v0.0.0' if no tag is found and exports the latest tag to an environment variable for further usage.