V2 (#55)
* Delete Docs

* Add Sphinx documentation setup

This commit introduces a new Sphinx documentation setup for the WrenchCL project. A configuration file, make file, an index, and a batch file were created to ensure efficient documentation process. The project's documentation now supports autodoc, viewcode, and napoleon extensions for better readability.

* Update GitHub Actions workflow for Python project

This commit fine-tunes multiple sections of our GitHub Actions workflow. Some redundant comments were removed for better readability while image versions for actions were updated to the latest. A critical permission block was added to resolve workflow dispatch issues. Additionally, the workflow now includes steps for generating Sphinx documentation and deploying it to GitHub Pages.

* Add Sphinx documentation generation in GitHub Actions

A new workflow file has been added to the .github directory that automatically generates Sphinx documentation. This workflow is triggered manually and deploys the built documentation to GitHub Pages. Moreover, it operates in the latest Ubuntu environment with Python 3.11.

* Enable publishing to Py

* Update pytest command in GitHub workflow

The pytest command in the GitHub actions workflow has been updated to include the coverage report in XML format. An additional parameter is also added to the Codecov action to specify the location of the coverage report file.

* Update pytest command and Codecov upload in GitHub workflow

The pytest command has been updated to correctly target the WrenchCL application for coverage report. The steps for uploading the coverage report to Codecov within the GitHub workflow were improved to remove redundant configurations and include verbose logging for better troubleshooting.

* Update pytest run command and relocate test files

The commit modified the command to execute pytest, explicitly setting it to run on the /tests directory. Also, it relocated the test modules from resources/tests/ to the tests/ directory, improving the project structure and organization.

* Update pytest run command and relocate test files

The commit modified the command to execute pytest, explicitly setting it to run on the /tests directory. Also, it relocated the test modules from resources/tests/ to the tests/ directory, improving the project structure and organization.

* Update pytest command in GitHub workflows

The pytest command in the GitHub workflow has been simplified. The target test directory "./tests" has been placed after the "--cov=WrenchCL" argument for clarity. Additionally, the verbose mode previously used with the Codecov action is turned off to minimize output verbosity.

* Optimize pytest command and remove Codecov action

The commit simplifies the pytest command in the GitHub workflows configuration file, making the target test directory more explicit. Additionally, the Codecov

* Update file path in Sphinx documentation generation

The commit updates the file path used in the command for generating Sphinx documentation. Specifically, it changes the file from `config.py` to `conf.py` to match the correct file for configuring the Sphinx documentation generation.

* Update file path in Sphinx documentation generation

The commit updates the file path used in the command for generating Sphinx documentation. Specifically, it changes the file from `config.py` to `conf.py` to match the correct file for configuring the Sphinx documentation generation.

* Update Sphinx configuration and GitHub workflow

The html_static_path in Sphinx configuration has been updated for clarity. Additionally, sphinx-rtd-theme has been included in the GitHub actions workflow. These changes ensure the Sphinx documentation generation is set correctly and theme dependency is properly handled in CI workflow.

* Add .nojekyll file to Sphinx build workflow

This commit adds a .nojekyll file to the Sphinx build in the GitHub workflows for both publishing to PyPi and conducting Sphinx tests. It also updates the deployment step to exclude .nojekyll during cleanup, ensuring the file remains in the deployed gh-pages branch.

* Separate documentation generation and deployment from PyPi publishing

The code related to Sphinx documentation generation and deployment to GitHub Pages has been moved from the 'publish-to-pypi' workflow to a separate workflow named 'generate-docs'. This separation makes each workflow specific to a single responsibility, simplifies the 'publish-to-pypi' workflow, and provides greater flexibility in controlling when and how documentation is built and deployed.

* Separate documentation generation and deployment from PyPi publishing

The code related to Sphinx documentation generation and deployment to GitHub Pages has been moved from the 'publish-to-pypi' workflow to a separate workflow named 'generate-docs'. This separation makes each workflow specific to a single responsibility, simplifies the 'publish-to-pypi' workflow, and provides greater flexibility in controlling when and how documentation is built and deployed.

* Add DD_CIVISIBILITY_ITR_ENABLED parameter in workflow file

A new parameter named DD_CIVISIBILITY_ITR_ENABLED has been added into the `.github/workflows/publish-to-pypi.yml` workflow file. Its default value has been set as false. This configuration change affects the visibility settings of the CI/CD process.

* Remove Sphinx documentation test workflow

This commit removes the .github/workflows/sphinx-test.yml file. The workflow was previously used to test Sphinx documentation generation and deployment but it is no longer required. This leads to a cleaner, leaner codebase and less unnecessary CI/CD activities.

* Remove unused environment variable from workflow

The environment variable `DD_CIVISIBILITY_ITR_ENABLED` was removed from the GitHub Actions workflow for publishing to PyPi. This variable wasn't being used in the process, improving the readability and maintainability of the workflow configuration.

* Add pytestmark to skip tests conditionally

A pytestmark has been added to the tests in `tests/test_imports.py` and `tests/test_tools.py`. This decorator flags these tests as unskippable, providing context (datadog_itr_unskippable) as a reason. This will ensure these critical tests are always executed regardless of any conditions.

* Add shutdown logging function at exit

In the files 'test_imports.py' and 'test_tools.py' a function was added to shutdown the logging system when the program exits. This is done using the 'atexit' module. This change ensures that all logs are properly written and resources are cleaned up.

* Add fetch for all tags in GitHub Actions

The update ensures that all Git tags are fetched before trying to get the latest version tag in the GitHub Actions workflow for generating documentation. It also updates the fall back version to 'v0.0.0' if no tag is found and exports the latest tag to an environment variable for further usage.