V2 (#59)
* Add coverage to pytest and remove doc generation step

This commit introduces pytest code coverage for better insights into test execution. This is accomplished through the use of 'coverage' and 'pytest-cov' tools. Moreover, Sphinx documentation generation and the GitHub Pages deployment steps were removed from the workflow.

* Update documentation home page

Added a title, a centered WrenchCL logo, some relevant badges, and a short description about the WrenchCL library to the documentation home page. This should provide visitors a quick overview of the project.

* Refactor test imports and enhance workflow CI/CD

In this commit, the 'test_tools.py' file has been refactored to enhance readability by separating import statements. Moreover, the CI/CD workflow on GitHub Actions has been improved to handle pull requests on branch 'v2'. It now runs tests in

* Add coverage directory creation to GitHub workflow

Added lines of code in the GitHub workflow .yml file to create a 'coverage' directory and move the lcov.info file into this directory before uploading the coverage report. This will help organize the generated test coverage reports better.

* Simplify pytest command in GitHub workflow

The pytest command in the GitHub workflow file, "publish-to-pypi.yml", has been simplified. Extra reports generation like xml and html has been removed, focusing only on lcov coverage reports. Additionally, the creation of the coverage directory and the move of lcov.info file operations have been removed, as they're not necessary.

* Remove coverage report posting in GitHub workflow

The commit eliminates the step of posting a coverage report in the GitHub Actions workflow for publishing to Pypi. This was done to streamline the workflow and avoid clutter of coverage reports in pull request comments.

* Add documentation link to README

A link to the project documentation on Read the Docs was added to the README file. This will make it easier for users to get detailed information and instructions about the project.